+++ src/commands/create.js
@@ -6,8 +6,6 @@ const fsExtra = require( 'fs-extra' );
 const sudo = require( 'sudo-prompt' );
 const compose = require( 'docker-compose' );
 const which = require( 'which' );
-const mkcert = require( 'mkcert-prebuilt' );
-const shellEscape = require( 'shell-escape' );
 
 const { startGlobal } = require( '../gateway' );
 const environment = require( '../environment' );
@@ -44,7 +42,7 @@ async function createCommand( spinner, defaults = {} ) {
     await saveYaml( 'wp-cli.yml', { ssh: 'docker-compose:phpfpm' } );
 
     await makeCopyConfigs( spinner, fsExtra )( paths, answers );
-    await makeCert( spinner, mkcert, shellEscape )( envSlug, envHosts );
+    await makeCert( spinner )( envSlug, envHosts );
 
     await startGlobal( spinner );
     await makeDatabase( spinner )( envSlug );